requirements.txt
blinker            1.8.1
certifi            2023.5.7
charset-normalizer 3.1.0
click              8.1.7
colorama           0.4.6
contourpy          1.1.1
cycler             0.12.1
Flask              3.0.3
fonttools          4.43.1
idna               3.4
itsdangerous       2.2.0
Jinja2             3.1.3
kiwisolver         1.4.5
MarkupSafe         2.1.5
matplotlib         3.8.0
numpy              1.26.1
packaging          23.2
pandas             2.2.1
Pillow             10.1.0
pyparsing          3.1.1
python-dateutil    2.8.2
pytz               2024.1
requests           2.31.0
scipy              1.11.3
six                1.16.0
SpeechRecognition  3.10.0
tzdata             2024.1
urllib3            2.0.3
Werkzeug           3.0.2
gunicorn